Support for Clang 18 and GCC 13
Deprecate TreeNOTOPEN in favor of TreeNOT_OPEN
Increase test timeout to 1h
Add opts for Ubuntu 24

+++ xmdsshr/XmdsWaveform.c
@@ -1816,7 +1816,7 @@ static void AutoScale(int num, float *value, float *minval, float *maxval,
   vmax = max(vmin, min(vmax, huge));
   if (vmax == vmin)
   {
-    if (vmin == huge)
+    if (vmin == (float)huge)
       vmin = vmax = 0.0;
     span = fabs(vmax);
     if (span <= 0.0)
